rename: 최상위 api 폴더 apis 로 이름 변경
소셜 로그인 관련 오류 처리
